+++ src/main/java/fi/dy/masa/malilib/compat/modmenu/ModMenuImpl.java
@@ -1,13 +1,13 @@
 package fi.dy.masa.malilib.compat.modmenu;
 
 import fi.dy.masa.malilib.MaLiLibConfigGui;
-import org.thinkingstudio.mafglib.loader.gui.ModConfigScreenFactory;
-import org.thinkingstudio.mafglib.loader.gui.ModConfigScreenInitializer;
+import net.neoforged.neoforge.client.gui.IConfigScreenFactory;
+import org.thinkingstudio.mafglib.loader.entrypoints.ConfigScreenEntrypoint;
 
-public class ModMenuImpl implements ModConfigScreenInitializer
+public class ModMenuImpl implements ConfigScreenEntrypoint
 {
     @Override
-    public ModConfigScreenFactory getModConfigScreenFactory()
+    public IConfigScreenFactory getModConfigScreenFactory()
     {
         return (modContainer, screen) -> {
             MaLiLibConfigGui gui = new MaLiLibConfigGui();