+++ src/main/java/fi/dy/masa/malilib/datagen/MaLiLibDataGen.java
@@ -1,10 +1,12 @@
 package fi.dy.masa.malilib.datagen;
 
 import net.neoforged.neoforge.data.event.GatherDataEvent;
+import org.thinkingstudio.mafglib.loader.entrypoints.DataGeneratorEntrypoint;
 
-public class MaLiLibDataGen
+public class MaLiLibDataGen implements DataGeneratorEntrypoint
 {
-    public static void onInitializeDataGenerator(GatherDataEvent.Client event)
+    @Override
+    public void onInitializeDataGenerator(GatherDataEvent.Client event)
     {
         var generator = event.getGenerator();
         var output = generator.getPackOutput();