+++ src/main/java/org/thinkingstudio/mafglib/loader/entrypoints/ConfigScreenEntrypoint.java
@@ -1,0 +1,8 @@
+package org.thinkingstudio.mafglib.loader.entrypoints;
+
+import net.neoforged.fml.IExtensionPoint;
+import net.neoforged.neoforge.client.gui.IConfigScreenFactory;
+
+public interface ConfigScreenEntrypoint extends IExtensionPoint {
+    IConfigScreenFactory getModConfigScreenFactory();
+}